requirements.txt
Flask
Flask-Limiter
firebase-admin
google-generativeai
google-cloud-firestore
google-cloud-core
google-auth
requests
python-dotenv
gunicorn
edge-tts
pydub
feedparser
beautifulsoup4
html5lib
nltk
aiofiles
redis==4.5.4
dramatiq[redis]==1.18.0